+++ src/idf-mobilite.js
@@ -12,6 +12,7 @@ import { renderBUS } from "./render/render-bus.js";
 import { renderRER } from "./render/render-rer.js";
 import { renderMessages } from "./render/render-message.js";
 import { IDFMobiliteStyles } from "./render/render-style.js";
+import { images } from "./images/index.js";
 
 class IDFMobiliteCard extends LitElement {
     static get properties() {
@@ -31,7 +32,6 @@ class IDFMobiliteCard extends LitElement {
         if (!this.config || !this.hass) {
             return html``;
         }
-        const imagesUrl = 'https://raw.githubusercontent.com/bontiv/lovelace-idf-mobilite/main/src/images/'
         return html`
         <ha-card>
             <div class="${this.getBorderClass()}">
@@ -98,7 +98,7 @@ class IDFMobiliteCard extends LitElement {
                     ${this.config.show_screen
                     ? html`
                         <div class="ratp-img">
-                            <img src="${imagesUrl}ratp.png" class="ratp-image">
+            <img src="${images['ratp.png']}" class="ratp-image">
                             <div class="blink-point"></div>
                         </div>
                         `
@@ -138,9 +138,8 @@ class IDFMobiliteCard extends LitElement {
         const model = parseRerFromSiri(lineDatas, exclude_lines, exclude_lines_ref, nb_departure, groupDestination, groupDestinationName, max_delay, show_replacement_bus)
         if (!model)
             return
-        const imagesUrl = 'https://raw.githubusercontent.com/bontiv/lovelace-idf-mobilite/main/src/images/'
          return html`
-            ${renderRER(model, this.config, imagesUrl, second_entity)}
+            ${renderRER(model, this.config, images, second_entity)}
         `;
     }
     /********************************************************
@@ -152,9 +151,8 @@ class IDFMobiliteCard extends LitElement {
         const model = parseBusFromSiri(lineDatas, exclude_lines, exclude_lines_ref, included_destination, show_only_included)
         if (!model)
             return
-        const imagesUrl = 'https://raw.githubusercontent.com/bontiv/lovelace-idf-mobilite/main/src/images/'
         return html`
-            ${renderBUS(model, this.config, imagesUrl, second_entity)}
+            ${renderBUS(model, this.config, images, second_entity)}
         `;
     }
 
@@ -173,9 +171,8 @@ class IDFMobiliteCard extends LitElement {
             messagesList.map(id => this.hass.states[id]).filter(Boolean)
         );
 
-        const imagesUrl = 'https://raw.githubusercontent.com/bontiv/lovelace-idf-mobilite/main/src/images/';
         const noScroll = this.config.no_messages_scroll === true;
-        const htmlString = renderMessages(model, this.config, imagesUrl, noScroll);
+        const htmlString = renderMessages(model, this.config, images, noScroll);
 
         if (!htmlString) {
             return html``;